Move pcr_beta_coef() function to regressors module - It didn't really make sense in the stats module since it's only   used with PCR.
