+++ tests/test_regressors.py
@@ -231,7 +231,7 @@ class TestPCRBetaCoef(unittest.TestCase):
         x_reduced = pcomp.transform(x_scaled)
         ols = linear_model.LinearRegression()
         ols.fit(x_reduced, y)
-        beta_coef = stats.pcr_beta_coef(ols, pcomp)
+        beta_coef = regressors.pcr_beta_coef(ols, pcomp)
         self.assertEqual(beta_coef.shape, ols.coef_.shape)
